refactor pages folder
