+++ frontend/src/components/pages/Modeler/ModelerSidebar/PropertiesPanel/PropertiesPanelSections/PPParameterSection/singleComponents/PPParameterInput.jsx
@@ -6,11 +6,11 @@ import {
   UnlockOutlined,
 } from '@ant-design/icons';
 import PropTypes from 'prop-types';
-import styles from '../../ModelerSidebar.module.css';
+import styles from '../../../../ModelerSidebar.module.css';
 import {
   setPropertyForParameter,
   parameterPropertyStatus,
-} from '../../../../../../utils/sessionStorageUtils/localSsotController/parameters';
+} from '../../../../../../../../utils/sessionStorageUtils/localSsotController/parameters';
 
 const { Text } = Typography;
