+++ frontend/src/components/pages/Modeler/ModelerSidebar/PropertiesPanel/PropertiesPanelSections/PPRpaSection/PPRpaSection.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Typography, Space } from 'antd';
 import PropTypes from 'prop-types';
-import PropertiesPanelApplicationDropdown from '../PropertiesPanelComponents/PPApplicationDropdown';
-import PropertiesPanelTaskDropdown from '../PropertiesPanelComponents/PPTaskDropdown';
+import PropertiesPanelApplicationDropdown from './singleComponents/PPApplicationDropdown';
+import PropertiesPanelTaskDropdown from './singleComponents/PPTaskDropdown';
 
 const { Text } = Typography;
